리펙토링: [FASTapi] exception
- Add exception
- custom exception handler

+++ fastapi-1/db/db_user.py
@@ -1,3 +1,5 @@
+from fastapi import HTTPException, status
+
 from db.hash import Hash
 from sqlalchemy.orm.session import Session
 from schemas import UserBase
@@ -21,11 +23,19 @@ def get_all_users(db: Session):
 
 
 def get_user(db: Session, id: int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
+    return user
 
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
+
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -37,6 +47,10 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
+    
     db.delete(user)
     db.commit()
     return 'ok'
